# docs/swagger.yaml
basePath: /
definitions:
...
  errors.ErrorValidation:
    properties:
      field:
        type: string
      reason:
        type: string
      reason_value:
        type: string
    type: object
...
  model.Response:
    properties:
      message:
        type: string
    type: object
  model.ResponseWithError:
    properties:
      error:
        items:
          $ref: '#/definitions/errors.ErrorValidation'
        type: array
      message:
        type: string
    type: object
  model.Tag:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  tag.TagRequest:
    properties:
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
  tag.TagResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      message:
        type: string
    type: object
  user.UserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  user.UserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.User'
        type: array
      message:
        type: string
    type: object
host: localhost
info:
  contact: {}
  title: Event Management
  version: "1.0"
paths:
...
  /tag:
    get:
      consumes:
      - application/json
      description: Find all Tag with their data
      operationId: tag-find-all
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Find All Tag
      tags:
      - Tag
    post:
      consumes:
      - application/json
      description: Create a Tag
      operationId: tag-create
      parameters:
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/tag.TagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Create a Tag
      tags:
      - Tag
  /tag/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Tag with their data
      operationId: tag-delete
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Delete a Tag
      tags:
      - Tag
    patch:
      consumes:
      - application/json
      description: Update a Tag with their data
      operationId: tag-update
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/tag.TagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Update a Tag
      tags:
      - Tag
  /user:
    get:
      consumes:
      - application/json
      description: Find all user with their data
      operationId: user-find-all
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Find All User
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create a User
      operationId: user-create
      parameters:
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/user.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Create a User
      tags:
      - User
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user with their data
      operationId: user-delete
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Delete a User
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Update a user with their data
      operationId: user-update
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/user.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ResponseWithError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiBearerToken: []
      summary: Update a User
      tags:
      - User
securityDefinitions:
  ApiBearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"